+++ lib/flutter_loading_animation_kit.dart
@@ -1,3 +1,6 @@
 library flutter_loading_animation_kit;
 
 export 'src/four_circle_pulse.dart';
+export 'src/yin_and_yang.dart';
+export 'src/ripple_ring.dart';
+export 'src/line_ellipsis.dart';